Removed workgroup templates on opalgobase classes
